+++ docs/examples/datastructures/pagination/using_offset_pagination.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 from pydantic_factories import ModelFactory
 
 from starlite import Starlite, get
-from starlite.utils.pagination import AbstractSyncOffsetPaginator, OffsetPagination
+from starlite.pagination import AbstractSyncOffsetPaginator, OffsetPagination
 
 
 class Person(BaseModel):
