+++ docs/lib/reference/pagination.rst
@@ -1,0 +1,5 @@
+pagination
+----------
+
+.. automodule:: starlite.pagination
+   :members:
