+++ docs/lib/reference/utils/index.rst
@@ -4,7 +4,6 @@ starlite.utils
 .. toctree::
     :maxdepth: 1
 
-    pagination
     predicates
     scope
     sync
