--- docs/lib/reference/utils/pagination.rst
@@ -1,5 +1,0 @@
-pagination
-----------
-
-.. automodule:: starlite.utils.pagination
-   :members:
