+++ tests/test_pagination.py
@@ -5,9 +5,7 @@ import pytest
 
 from starlite import get
 from starlite.app import DEFAULT_OPENAPI_CONFIG
-from starlite.status_codes import HTTP_200_OK
-from starlite.testing import create_test_client
-from starlite.utils.pagination import (
+from starlite.pagination import (
     AbstractAsyncClassicPaginator,
     AbstractAsyncCursorPaginator,
     AbstractAsyncOffsetPaginator,
@@ -18,6 +16,8 @@ from starlite.utils.pagination import (
     CursorPagination,
     OffsetPagination,
 )
+from starlite.status_codes import HTTP_200_OK
+from starlite.testing import create_test_client
 from tests import Person, PersonFactory
