+++ tests/utils/test_predicates.py
@@ -21,8 +21,8 @@ from typing import (
 import pytest
 
 from starlite import Response, get
+from starlite.pagination import CursorPagination
 from starlite.utils import is_any, is_class_and_subclass, is_optional_union, is_union
-from starlite.utils.pagination import CursorPagination
 from starlite.utils.predicates import (
     is_generic,
     is_mapping,
